+++ include/aligator/modelling/centroidal/centroidal-wrapper.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "aligator/core/unary-function.hpp"
-#include <proxsuite-nlp/third-party/polymorphic_cxx14.hpp>
+#include "aligator/third-party/polymorphic_cxx14.h"
 
 namespace aligator {