+++ include/aligator/modelling/costs/quad-state-cost.hpp
@@ -1,6 +1,6 @@
 /// @file
 /// Convenience classes to define quadratic state or control cost functions.
-/// @copyright Copyright (C) 2023 LAAS-CNRS, INRIA
+/// @copyright Copyright (C) 2023 LAAS-CNRS, 2023-2025 INRIA
 #pragma once
 
 #include "aligator/modelling/state-error.hpp"
@@ -15,14 +15,23 @@ struct QuadraticStateCostTpl : QuadraticResidualCostTpl<Scalar> {
   using Base = QuadraticResidualCostTpl<Scalar>;
   using StateError = StateErrorResidualTpl<Scalar>;
   using Manifold = ManifoldAbstractTpl<Scalar>;
+  using StageFunction = StageFunctionTpl<Scalar>;
 
   // StateError's space variable holds a pointer to the state manifold
-  QuadraticStateCostTpl(StateError resdl, const MatrixXs &weights)
-      : Base(resdl.space_, resdl, weights) {}
+  QuadraticStateCostTpl(const StateError &resdl, const MatrixXs &weights)
+      : Base(resdl.space_, xyz::polymorphic<StageFunction>(resdl), weights) {}
 
   QuadraticStateCostTpl(xyz::polymorphic<Manifold> space, const int nu,
                         const ConstVectorRef &target, const MatrixXs &weights)
-      : QuadraticStateCostTpl(StateError(space, nu, target), weights) {}
+      : QuadraticStateCostTpl(StateError(std::move(space), nu, target),
+                              weights) {}
+
+  template <typename U,
+            typename = std::enable_if_t<!is_polymorphic_of_v<Manifold, U>>>
+  QuadraticStateCostTpl(U &&space, const int nu, const ConstVectorRef &target,
+                        const MatrixXs &weights)
+      : QuadraticStateCostTpl(StateError(std::forward<U>(space), nu, target),
+                              weights) {}
 
   void setTarget(const ConstVectorRef target) { residual().target_ = target; }
   ConstVectorRef getTarget() const { return residual().target_; }
@@ -39,28 +48,34 @@ struct QuadraticControlCostTpl : QuadraticResidualCostTpl<Scalar> {
   ALIGATOR_DYNAMIC_TYPEDEFS(Scalar);
   using Base = QuadraticResidualCostTpl<Scalar>;
   using Manifold = ManifoldAbstractTpl<Scalar>;
-  using Error = ControlErrorResidualTpl<Scalar>;
+  using ControlError = ControlErrorResidualTpl<Scalar>;
+  using StageFunction = StageFunctionTpl<Scalar>;
 
-  QuadraticControlCostTpl(xyz::polymorphic<Manifold> space, Error resdl,
-                          const MatrixXs &weights)
-      : Base(space, resdl, weights) {}
+  QuadraticControlCostTpl(xyz::polymorphic<Manifold> space,
+                          const ControlError &resdl, const MatrixXs &weights)
+      : Base(std::move(space), xyz::polymorphic<StageFunction>(resdl),
+             weights) {}
 
   QuadraticControlCostTpl(xyz::polymorphic<Manifold> space, int nu,
                           const ConstMatrixRef &weights)
-      : QuadraticControlCostTpl(space, Error(space->ndx(), nu), weights) {}
+      : QuadraticControlCostTpl(space, ControlError(space->ndx(), nu),
+                                weights) {}
 
   QuadraticControlCostTpl(xyz::polymorphic<Manifold> space,
                           const ConstVectorRef &target,
                           const ConstMatrixRef &weights)
-      : QuadraticControlCostTpl(space, Error(space->ndx(), target), weights) {}
+      : QuadraticControlCostTpl(space, ControlError(space->ndx(), target),
+                                weights) {}
 
   void setTarget(const ConstVectorRef &target) { residual().target_ = target; }
   ConstVectorRef getTarget() const { return residual().target_; }
 
 protected:
-  Error &residual() { return static_cast<Error &>(*this->residual_); }
-  const Error &residual() const {
-    return static_cast<const Error &>(*this->residual_);
+  ControlError &residual() {
+    return static_cast<ControlError &>(*this->residual_);
+  }
+  const ControlError &residual() const {
+    return static_cast<const ControlError &>(*this->residual_);
   }
 };