+++ include/aligator/modelling/multibody/frame-collision.hpp
@@ -8,7 +8,7 @@
 #include <pinocchio/multibody/frame.hpp>
 #include <pinocchio/algorithm/geometry.hpp>
 
-#include <proxsuite-nlp/third-party/polymorphic_cxx14.hpp>
+#include "aligator/third-party/polymorphic_cxx14.h"
 
 namespace aligator {