+++ include/aligator/modelling/spaces/cartesian-product.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "aligator/core/manifold-base.hpp"
-#include "proxsuite-nlp/third-party/polymorphic_cxx14.hpp"
+#include "aligator/third-party/polymorphic_cxx14.h"
 
 #include <type_traits>